refactor super_docopt to mister_bin
