+++ lib/sla.rb
@@ -2,7 +2,7 @@ require 'webcache'
 require 'nokogiri'
 require 'colsole'
 require 'singleton'
-require 'super_docopt'
+require 'mister_bin'
 require 'uri'
 
 require 'sla/version'
@@ -14,7 +14,7 @@ require 'sla/formatters/base'
 require 'sla/formatters/verbose'
 require 'sla/formatters/tty'
 
-require 'sla/command_line'
+require 'sla/command'
 
 if ENV['BYEBUG']
   # :nocov:
