+++ sla.gemspec
@@ -17,8 +17,8 @@ Gem::Specification.new do |s|
   s.license     = 'MIT'
   s.required_ruby_version = ">= 2.4.0"
 
-  s.add_dependency 'colsole', '~> 0.7'
-  s.add_dependency 'super_docopt', '~> 0.1'
-  s.add_dependency 'webcache', '~> 0.4'
-  s.add_dependency 'nokogiri', '~> 1.8'
+  s.add_dependency 'colsole', '~> 0.7' #, '>= 0.7.1'
+  s.add_dependency 'mister_bin', '~> 0.7'
+  s.add_dependency 'webcache', '~> 0.5.1'
+  s.add_dependency 'nokogiri', '~> 1.10'
 end
